fix: add timeout for registrar service - use 30 second timeout for registrar and verifier
Cause: The registrar service can fail, but the Ansible service reports success.

Consequence: The user does not know that the registrar was mis-configured.

Fix: Check the registrar ip and port to ensure they become available.  Use a 30
second timeout for both the verifier and registrar as the default 300 seconds for
the `wait_for` module is too long.

Result: Users are notified in a timely manner if the services failed to start.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>