ci(packit): fix extra parameters for TF (#75)
* ci(tmt): use tmt's feature/epel for enabling EPEL

Signed-off-by: Matej Focko <mfocko@redhat.com>

* ci(tmt): force EPEL before installing deps

Essential dependencies are handled with `order = 30`, therefore it is
needed to adjust the priority of enabling the EPEL, otherwise the deps
(even from EPEL) are installed before the repo gets enabled.

For more details see discussion here:

    https://redhat-internal.slack.com/archives/C04LRPNDZ4Y/p1752748187699709

Suggested-by: Petr Šplíchal <psplicha@redhat.com>
Signed-off-by: Matej Focko <mfocko@redhat.com>

* ci(packit): run only `packit-ci` plan via Packit

The other plan is triggered externally and doesn't work correctly when
triggered via Packit.

Signed-off-by: Matej Focko <mfocko@redhat.com>

---------

Signed-off-by: Matej Focko <mfocko@redhat.com>